// HrmsAngular/src/app/models/Deduction/all-deduction.model.ts
import { DriverAllowace } from "../Addition/driver-addition-allowance.model";

export class AllDeduction {
  id: number;
  selectedDriver: any[];
  empID: number;
  structureID: number;
  salaryHeadID: number;
  salaryHeadType: number;
  yearID: number;
  taxYearID: number;
  amount: number;
  salaryTypeID: number;
  basedOnID: number;
  createdDate: string;
  periodID: number;
  sortOrder: number;
  periodName: string;
  grade: number;
  companyID: number;
  days: number;
  empName: string;
  designation: string;
  department: string;
  empCode: string;
  userTypeID:number;
}